+++ bus_user/__init__.py
@@ -1,9 +1,9 @@
 from .history import *
-from .serial_user import *
-from .serial_emulator import *
+from .serial_client import *
+from .serial_server import *
 
 # from .bus_i2c import *  # import only Linux supported modul return exx
 try:
-    from .i2c_user import *    # import only Linux supported modul return exx
+    from .i2c_client import *    # import only Linux supported modul return exx
 except:
     pass